requirements.txt
Flask
gunicorn
Jinja2
jsonschema
numpy
pandas
plotly
requests
urllib3
Werkzeug
statsmodels
email_validator
python-dotenv
Flask-SQLAlchemy
Flask-WTF
Flask-Login
pymysql
Flask-Mail